requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
faiss-cpu==1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
pydantic==2.5.0
pyyaml==6.0.1
lxml==4.9.3
